requirements_streamlit.txt
streamlit>=1.46.0
langchain>=0.3.24
openai>=1.75.0
faiss-cpu>=1.11.0
sentence-transformers>=4.1.0
python-dotenv>=1.1.0
plotly>=6.2.0
pandas>=2.3.0
requests>=2.32.0
beautifulsoup4>=4.13.0